notes api

+++ app/Http/Controllers/NoteController.php
@@ -6,8 +6,10 @@ use App\Http\Requests\StoreNoteRequest;
 use App\Http\Requests\UpdateNoteRequest;
 use App\Http\Resources\NoteResource;
 use App\Models\Note;
+use App\Models\Organization;
+use App\Models\User;
 use App\Repositories\NoteRepository;
-use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Response;
@@ -21,24 +23,28 @@ class NoteController extends Controller
         $this->noteRepository = $noteRepository;
     }
 
-    public function index(Request $request): AnonymousResourceCollection
+    public function index(Request $request, Organization $organization): AnonymousResourceCollection
     {
-        // abort_if($request->user()->cannot('viewAny', Note::class), 403);
+        $this->authorize($request->user(), 'viewAny', $organization, Note::class);
 
-        $notes = $this->noteRepository->getNotes($request->all());
+        $notes = $this->noteRepository->getNotes($organization, $request->all());
 
         return NoteResource::collection($notes);
     }
 
-    public function store(StoreNoteRequest $request): NoteResource | RedirectResponse
+    public function store(StoreNoteRequest $request, Organization $organization): NoteResource|JsonResponse
     {
+        $this->authorize($request->user(), 'create', $organization, Note::class);
+
         try {
-            $note = $this->noteRepository->insert($request->all());
+            /* @phpstan-ignore-next-line */
+            $note = $this->noteRepository->insert($organization, $request->user(), $request->all());
 
             return new NoteResource($note);
         } catch (\Throwable $th) {
-            return back()->with('error', $th->getMessage());
+            return response()->json(['error' => $th->getMessage()], 422);
         }
+
     }
 
     public function show(Note $note): void
@@ -46,25 +52,37 @@ class NoteController extends Controller
         //
     }
 
-    public function update(UpdateNoteRequest $request, Note $note): NoteResource | RedirectResponse
+    public function update(UpdateNoteRequest $request, Organization $organization, Note $note): NoteResource|JsonResponse
     {
+        $this->authorize($request->user(), 'update', $organization, $note);
+
         try {
             $note = $this->noteRepository->update($note, $request->all());
 
             return new NoteResource($note);
         } catch (\Throwable $th) {
-            return back()->with('error', $th->getMessage());
+            return response()->json(['error' => $th->getMessage()], 422);
         }
     }
 
-    public function destroy(Note $note): Response | RedirectResponse
+    public function destroy(Request $request, Organization $organization, Note $note): Response|JsonResponse
     {
+        $this->authorize($request->user(), 'delete', $organization, $note);
+
         try {
             $this->noteRepository->delete($note);
 
             return response()->noContent();
         } catch (\Throwable $th) {
-            return back()->with('error', $th->getMessage());
+            return response()->json(['error' => $th->getMessage()], 422);
         }
+    }
+
+    public function authorize(?User $user, string $action, Organization $organization, mixed $classOrModel): void
+    {
+        if ($user && $user->cannot($action, [$classOrModel, $organization])) {
+            abort(403);
+        }
+
     }
 }
